Se agrega la validacion para evitar la eliminacion de roles en uso y tambien se agregaron las tablas de estado (State), municipio (Municipality), parroquia (Parish) con la relacion de estado-municipio (1:n), minicipio-parroquia (1:n)

+++ backend/src/controllers/Role.ts
@@ -89,6 +89,11 @@ export const deleteRole = async (id: number): Promise<boolean> => {
             return false;
         }
 
+        const countUsers = await User.count({ where: { roleId: id } });
+        if (countUsers > 0) {
+            throw new Error("No se puede eliminar el rol, ya que está en uso.");
+        }
+
         await Role.destroy({
             where: { id },
         });
